Fixing some rendering bugs.
The spec does not define the truthiness of zero, this
implementation still considers it False, but it now renders
as '0' if it is the value of a ValueBlock's tag.

Also adding support for looping over a section with
the list returned by a function call.

Test cases for the above are being added, and now pass.