Adding cystache_context_manager.
This lets all callbacks access the current context and template.
Because it is global there are threading issues, and enable_threading()
must be called if it is used in a multithreaded environment.

Also adding more tests for false value interpolation, "{{a}}" will
now render an empty string if a is False, None, [] or {}.